BotTipBooks/BotTipBooks/requirements.txt
botbuilder-integration-aiohttp>=4.11.0
botbuilder-dialogs>=4.11.0
botbuilder-ai>=4.11.0
datatypes-date-time>=1.0.0.a2
pyodbc>=4.0.30
pyadaptivecards>=0.1.1
azure-ai-textanalytics
botbuilder-schema>=4.11.0
botframework-connector>=4.11.0
